COG-30 CAP-API Retrieve a list of permissions
new version of cwdsDataModel,
Refactoring
