+++ src/main/java/gov/ca/cwds/idm/IdmResource.java
@@ -3,7 +3,6 @@ package gov.ca.cwds.idm;
 import gov.ca.cwds.idm.dto.User;
 import gov.ca.cwds.idm.service.DictionaryProvider;
 import gov.ca.cwds.idm.service.IdmService;
-import gov.ca.cwds.idm.service.PermissionService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import java.util.List;
@@ -20,10 +19,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(value = "/idm")
 public class IdmResource {
 
-  @Autowired private IdmService idmService;
+  @Autowired
+  private IdmService idmService;
 
-  @Autowired private PermissionService permissionService;
-  @Autowired private DictionaryProvider dictionaryProvider;
+  @Autowired
+  private DictionaryProvider dictionaryProvider;
 
   @RequestMapping(method = RequestMethod.GET, value = "/users", produces = "application/json")
   public List<User> getUsers() {
@@ -34,8 +34,8 @@ public class IdmResource {
   @ApiOperation(value = "Find User by ID", response = User.class)
   public ResponseEntity<User> getUser(
       @ApiParam(required = true, value = "The unique user ID", example = "userId1")
-          @PathVariable
-          @NotNull
+      @PathVariable
+      @NotNull
           String id) {
     return Optional.ofNullable(idmService.findUser(id))
         .map(user -> ResponseEntity.ok().body(user))
@@ -44,12 +44,10 @@ public class IdmResource {
 
   @RequestMapping(method = RequestMethod.GET, value = "/permissions", produces = "application/json")
   @ApiOperation(value = "Get List of possible permissions", response = List.class)
-  public List<String> getPermissions() {
-    return dictionaryProvider.getPermissions();
+  public ResponseEntity<List<String>> getPermissions() {
+    return Optional.ofNullable(dictionaryProvider.getPermissions())
+        .map(permissions -> ResponseEntity.ok().body(permissions)).orElseGet(() ->
+            ResponseEntity.notFound().build());
   }
 
-  @RequestMapping(method = RequestMethod.GET, value = "/permissions1", produces = "application/json")
-  public List<String> getPermissions1() {
-    return permissionService.getPermissionNames();
-  }
 }
